📝 Refactor QueryDsl, Chat
