+++ src/main/java/com/youprice/onion/service/product/impl/ProductServiceImpl.java
@@ -37,7 +37,6 @@ public class ProductServiceImpl implements ProductService {
     private final TownRepositoy townRepositoy;
     private final CategoryRepositoy categoryRepository;
     private final ProductRepository productRepository;
-    private final ProductRepository.Querydsl productRepositoryQuerydsl;
     private final ProductImageRepository productImageRepository;
     private final OrderRepository orderRepository;
 
@@ -60,7 +59,7 @@ public class ProductServiceImpl implements ProductService {
                 productRepository.save(product);
             }
         }
-        return productRepositoryQuerydsl.findAllBySearchRequirements(searchRequirements).map(ProductListDTO::new);
+        return productRepository.findAllBySearchRequirements(searchRequirements).map(ProductListDTO::new);
     }
 
 
@@ -291,7 +290,7 @@ public class ProductServiceImpl implements ProductService {
     //유저 판매 상품 목록
     @Override
     public Page<ProductSellListDTO> getProductSellListDTO(Long memberId, ProductProgress productProgress, Pageable pageable) {
-        return productRepositoryQuerydsl.findByMemberIdAndProductProgress(memberId, productProgress, pageable).map(ProductSellListDTO::new);
+        return productRepository.findByMemberIdAndProductProgress(memberId, productProgress, pageable).map(ProductSellListDTO::new);
     }
     //개인 유저 상품 리스트
     @Override
